+++ examples/helmholtz1D.py
@@ -6,15 +6,14 @@ import jax.numpy as jnp
 import matplotlib.pyplot as plt
 import sympy as sp
 
-from jaxfun.arguments import TestFunction, TrialFunction
-from jaxfun.basespace import n
+from jaxfun.galerkin.arguments import TestFunction, TrialFunction
 
 # from jaxfun.Jacobi import Jacobi as space
-from jaxfun.functionspace import FunctionSpace
-from jaxfun.inner import inner
-from jaxfun.Legendre import Legendre as space
+from jaxfun.galerkin.functionspace import FunctionSpace
+from jaxfun.galerkin.inner import inner
+from jaxfun.galerkin.Legendre import Legendre as space
 from jaxfun.operators import Div, Grad
-from jaxfun.utils.common import lambdify, ulp
+from jaxfun.utils.common import lambdify, n, ulp
 
 x = sp.Symbol("x", real=True)
 M = 50
