+++ examples/poisson1D_curv.py
@@ -25,14 +25,13 @@ import numpy as np
 import sympy as sp
 from sympy.plotting import plot3d_parametric_line
 
-from jaxfun.arguments import TestFunction, TrialFunction
-from jaxfun.basespace import n
 from jaxfun.coordinates import get_CoordSys
-from jaxfun.functionspace import FunctionSpace
-from jaxfun.inner import inner
-from jaxfun.Legendre import Legendre as space
+from jaxfun.galerkin.arguments import TestFunction, TrialFunction
+from jaxfun.galerkin.functionspace import FunctionSpace
+from jaxfun.galerkin.inner import inner
+from jaxfun.galerkin.Legendre import Legendre as space
 from jaxfun.operators import Div, Grad
-from jaxfun.utils.common import lambdify, ulp
+from jaxfun.utils.common import lambdify, n, ulp
 
 t = sp.Symbol("t", real=True)
 rv = (sp.sin(2 * sp.pi * t), sp.cos(2 * sp.pi * t), 2 * t)
