+++ examples/poisson3D.py
@@ -5,14 +5,13 @@ import jax.numpy as jnp
 from scipy import sparse as scipy_sparse
 from scipy.sparse import kron
 
-from jaxfun.arguments import TestFunction, TrialFunction
-from jaxfun.basespace import n
-from jaxfun.functionspace import FunctionSpace
-from jaxfun.inner import inner
-from jaxfun.Legendre import Legendre
+from jaxfun.galerkin.arguments import TestFunction, TrialFunction
+from jaxfun.galerkin.functionspace import FunctionSpace
+from jaxfun.galerkin.inner import inner
+from jaxfun.galerkin.Legendre import Legendre
+from jaxfun.galerkin.tensorproductspace import TensorProduct, tpmats_to_scipy_kron
 from jaxfun.operators import Div, Grad
-from jaxfun.tensorproductspace import TensorProduct, tpmats_to_scipy_kron
-from jaxfun.utils.common import lambdify, ulp
+from jaxfun.utils.common import lambdify, n, ulp
 
 M = 20
 bcs = {"left": {"D": 0}, "right": {"D": 0}}
