+++ src/jaxfun/galerkin/__init__.py
@@ -1,0 +1,18 @@
+from jaxfun.galerkin import (
+    Chebyshev as Chebyshev,
+    Fourier as Fourier,
+    Jacobi as Jacobi,
+    Legendre as Legendre,
+)
+from jaxfun.galerkin.arguments import (
+    JAXFunction as JAXFunction,
+    TestFunction as TestFunction,
+    TrialFunction as TrialFunction,
+)
+from jaxfun.galerkin.functionspace import FunctionSpace as FunctionSpace
+from jaxfun.galerkin.inner import inner as inner
+from jaxfun.galerkin.tensorproductspace import (
+    TensorProduct as TensorProduct,
+    TensorProductSpace as TensorProductSpace,
+    VectorTensorProductSpace as VectorTensorProductSpace,
+)
