+++ src/jaxfun/operators.py
@@ -3,13 +3,14 @@ Here we extend the Sympy operators Divergence, Gradient, Curl, Cross and Dot
 using curvilinear coordinates. The expressions used for computing the operators
 have been collected from the online book
 
-[1] Kelly, PA. Mechanics Lecture Notes: An introduction to Solid Mechanics.  
+[1] Kelly, PA. Mechanics Lecture Notes: An introduction to Solid Mechanics.
 Available from http://homepages.engineering.auckland.ac.nz/~pkel015/SolidMechanicsBooks/index.html
 
 """
 
 import collections
 from itertools import product
+from numbers import Number
 from typing import Any
 
 import numpy as np
@@ -17,16 +18,14 @@ import sympy as sp
 from sympy import Expr
 from sympy.core import preorder_traversal
 from sympy.core.add import Add
-from sympy.core.function import Derivative
-from sympy.core.function import diff as df
+from sympy.core.function import Derivative, diff as df
 from sympy.core.mul import Mul
 from sympy.printing.latex import LatexPrinter
 from sympy.printing.pretty.stringpict import prettyForm
-from sympy.vector import Dot as sympy_Dot
-from sympy.vector import VectorAdd, VectorMul, VectorZero
+from sympy.vector import Dot as sympy_Dot, VectorAdd, VectorMul, VectorZero
 from sympy.vector.dyadic import Dyadic, DyadicAdd, DyadicMul, DyadicZero
-from sympy.vector.operators import Curl as sympy_Curl
 from sympy.vector.operators import (
+    Curl as sympy_Curl,
     Divergence,
     Gradient,
 )
@@ -672,7 +671,7 @@ class Cross(Expr):
     Examples
     ========
 
-    >>> from jaxfun.arguments import CartCoordSys, x, y, z
+    >>> from jaxfun.coordinates import CartCoordSys, x, y, z
     >>> from jaxfun.operators import Cross
     >>> N = CartCoordSys("N", (x, y, z))
     >>> v1 = N.i + N.j + N.k
@@ -703,7 +702,7 @@ class Outer(Expr):
     Examples
     ========
 
-    >>> from jaxfun.arguments import CartCoordSys, x, y
+    >>> from jaxfun.coordinates import CartCoordSys, x, y
     >>> from jaxfun.operators import Outer
     >>> N = CartCoordSys("N", (x, y))
     >>> v1 = N.i
@@ -735,15 +734,35 @@ class Outer(Expr):
 
 
 class Source(Expr):
-
     def __new__(cls, expr):
         expr = sp.sympify(expr)
         obj = Expr.__new__(cls, expr)
         obj._expr = expr
-        return obj 
-    
+        return obj
+
     def doit(self, **hints):
         return self._expr.doit(**hints)
+
+
+class Constant(sp.Symbol):
+    def __new__(cls, name: str, val: Number, **assumptions):
+        obj = super().__new__(cls, name, **assumptions)
+        obj.val = val
+        return obj
+
+    def doit(self) -> Number:
+        return self.val
+
+
+class Identity(sp.Expr):
+    def __init__(self, sys: CoordSys):
+        self.sys = sys
+
+    def doit(self):
+        return sum(
+            self.sys.base_dyadics()[:: self.sys.dims + 1], sp.vector.DyadicZero()
+        )
+
 
 def diff(self, *args, **kwargs):
     """
@@ -774,9 +793,9 @@ def diff(self, *args, **kwargs):
 ## -> z = 4*N.x*N.j + 4*N.y*N.i
 # z.is_Vector
 ## -> False
-# z is now a type Add and not VectorAdd as it should be. 
+# z is now a type Add and not VectorAdd as it should be.
 # Using the doit function below is a hack around it
-# from jaxfun.arguments import CartCoordSys, x, y, z
+# from jaxfun.coordinates import CartCoordSys, x, y, z
 # from jaxfun.operators import Grad
 # N = CartCoordSys("N", (x, y, z))
 # f = N.x*N.y
