+++ src/jaxfun/utils/common.py
@@ -90,7 +90,7 @@ def lambdify(
     cse: bool = False,
     doctring_limit: int = 1000,
 ) -> Callable[[Iterable[Array]], Array]:
-    from jaxfun.forms import get_system
+    from jaxfun.galerkin.forms import get_system
 
     system = get_system(expr)
     expr = system.expr_base_scalar_to_psi(expr)
